requirements.txt
matplotlib
pandas
numpy
pillow
torch
torchvision
